ai/bento/pyproject.toml
[project]
name = "bento"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "bentoml>=1.4",
    "bitsandbytes>=0.47.0",
    "causal-conv1d>=1.5",
    "datasets==3.6.0",
    "fastapi[standard]>=0.116.1",
    "flash-linear-attention",
    "pillow>=11.3.0",
    "qwen-vl-utils>=0.0.11",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.0",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers" }
